+++ cdp-agentkit-core/cdp_agentkit_core/actions/__init__.py
@@ -43,8 +43,16 @@ from cdp_agentkit_core.actions.transfer import (
     TransferInput,
     transfer,
 )
+from cdp_agentkit_core.actions.uniswap_v3.create_pool import (
+    UNISWAP_V3_CREATE_POOL_PROMPT,
+    UniswapV3CreatePoolInput,
+    uniswap_v3_create_pool,
+)
 
 __all__ = [
+    "UNISWAP_V3_CREATE_POOL_PROMPT",
+    "UniswapV3CreatePoolInput",
+    "uniswap_v3_create_pool",
     "DEPLOY_NFT_PROMPT",
     "DeployNftInput",
     "deploy_nft",