+++ cdp-langchain/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Added `uniswap_v3_create_pool` action to the cdp toolkit.
+
 ## [0.0.1] - 2024-11-04
 
 ### Added